requirements.txt
streamlit
pandas
scikit-learn
nltk
matplotlib
wordcloud
google-generativeai